Clean front-end project

+++ onlineshop4dvds_api/DTOs/ProductDetailDto.cs
@@ -7,7 +7,7 @@ public class ProductDetailDto
     public required decimal Price { get; set; }
     public required string Description { get; set; }
     public required ICollection<string> Genres { get; set; }
-    public required float Ratings { get; set; }
+    public required double Ratings { get; set; }
     public required int NumbersOfReviews { get; set; }
     public required int Stock { get; set; }
 }
